# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Reservas de Restaurantes",
        "description": "API para la gestión de clientes, mesas y reservas.",
        "contact": {
            "name": "Vladimir Faundez",
            "email": "v.faundezh@gmail.com"
        },
        "license": {
            "name": "Creative Commons Attribution-NonCommercial-NoDerivatives 4.0 International",
            "url": "https://creativecommons.org/licenses/by-nc-nd/4.0/legalcode"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor principal de la API"
        }
    ],
    "paths": {
...
        "/api/v1/tables": {
            "get": {
                "tags": [
                    "Tables"
                ],
                "summary": "Get all tables",
                "description": "Retrieve a list of all tables.",
                "operationId": "getAllTable",
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/TableSuccess"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthenticatedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tables"
                ],
                "summary": "Create a new table",
                "description": "Create a new table.",
                "operationId": "createTable",
                "requestBody": {
                    "$ref": "#/components/requestBodies/StoreTableRequest"
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/TableSuccessId"
                    },
                    "400": {
                        "$ref": "#/components/responses/TableInvalidRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthenticatedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/tables/{table}": {
            "get": {
                "tags": [
                    "Tables"
                ],
                "summary": "Get a table by ID",
                "description": "Retrieve a table by its ID.",
                "operationId": "getTableById",
                "parameters": [
                    {
                        "name": "table",
                        "in": "path",
                        "description": "ID of the table to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/TableSuccessId"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthenticatedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundError"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tables"
                ],
                "summary": "Update a table",
                "description": "Update a table by its ID.",
                "operationId": "updateTable",
                "parameters": [
                    {
                        "name": "table",
                        "in": "path",
                        "description": "ID of the table to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateTableRequest"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/TableSuccessId"
                    },
                    "400": {
                        "$ref": "#/components/responses/TableInvalidRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthenticatedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundError"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tables"
                ],
                "summary": "Delete a table",
                "description": "Delete a table by its ID.",
                "operationId": "deleteTable",
                "parameters": [
                    {
                        "name": "table",
                        "in": "path",
                        "description": "ID of the table to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Table deleted successfully"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthenticatedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundError"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/tables/availables": {
            "get": {
                "tags": [
                    "Tables"
                ],
                "summary": "Get available tables",
                "description": "Retrieve a list of available tables.",
                "operationId": "getAvailableTables",
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/TableSuccess"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthenticatedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/tables/{table}/release": {
            "patch": {
                "tags": [
                    "Tables"
                ],
                "summary": "Release a table",
                "description": "Release a table by its ID.",
                "operationId": "releaseTable",
                "parameters": [
                    {
                        "name": "table",
                        "in": "path",
                        "description": "ID of the table to release",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Table released successfully"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthenticatedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundError"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/tables/{table}/occupied": {
            "patch": {
                "tags": [
                    "Tables"
                ],
                "summary": "Occupied a table",
                "description": "Occupied a table by its ID.",
                "operationId": "occupiedTable",
                "parameters": [
                    {
                        "name": "table",
                        "in": "path",
                        "description": "ID of the table to occupied",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Table occupied successfully"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthenticatedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundError"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Table": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "is_available": {
                        "type": "boolean",
                        "example": true
                    },
                    "capacity": {
                        "type": "integer",
                        "example": 4
                    },
                    "location": {
                        "type": "string",
                        "example": "indoor"
                    }
                },
                "type": "object"
            },
...
            "UnauthenticatedResponse": {
                "required": [
                    "status",
                    "message",
                    "data"
                ],
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated"
                    },
                    "data": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "NotFoundResponse": {
                "required": [
                    "status",
                    "message",
                    "data"
                ],
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Resource not found"
                    },
                    "data": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UnauthorizedResponse": {
                "required": [
                    "status",
                    "message",
                    "data"
                ],
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Unauthorized"
                    },
                    "data": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ServerErrorResponse": {
                "required": [
                    "status",
                    "message",
                    "data"
                ],
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Server error"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "Internal server error"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        },
        "responses": {
...
            "NotFoundError": {
                "description": "Reservation not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/NotFoundResponse"
                        }
                    }
                }
            },
            "UnauthenticatedError": {
                "description": "Unauthenticated user",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/UnauthenticatedResponse"
                        }
                    }
                }
            },
            "UnauthorizedError": {
                "description": "Unauthorized action",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/UnauthorizedResponse"
                        }
                    }
                }
            },
            "ServerError": {
                "description": "Internal server error",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ServerErrorResponse"
                        }
                    }
                }
            },
...
            "TableSuccess": {
                "description": "Successful table operation",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": true
                                },
                                "message": {
                                    "type": "string",
                                    "example": ""
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Table"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "TableSuccessId": {
                "description": "Successful table operation",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": true
                                },
                                "message": {
                                    "type": "string",
                                    "example": ""
                                },
                                "data": {
                                    "$ref": "#/components/schemas/Table"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "TableInvalidRequest": {
                "description": "Invalid request data",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Invalid request"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "field": {
                                                "type": "string",
                                                "example": "status"
                                            },
                                            "message": {
                                                "type": "string",
                                                "example": "The status field is required."
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
...
        "requestBodies": {
...
            "StoreTableRequest": {
                "description": "Data required to store a table",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "is_available",
                                "capacity",
                                "location"
                            ],
                            "properties": {
                                "is_available": {
                                    "type": "boolean",
                                    "example": true
                                },
                                "capacity": {
                                    "type": "integer",
                                    "example": 4
                                },
                                "location": {
                                    "type": "string",
                                    "example": "indoor"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "UpdateTableRequest": {
                "description": "Data required to update a table",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "is_available",
                                "capacity",
                                "location"
                            ],
                            "properties": {
                                "is_available": {
                                    "type": "boolean",
                                    "example": true
                                },
                                "capacity": {
                                    "type": "integer",
                                    "example": 4
                                },
                                "location": {
                                    "type": "string",
                                    "example": "indoor"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "JWT Authorization header usando el esquema Bearer.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}